Add files via upload
This is a simple calculator project that runs on the web. It is my second completed project in my web development journey. I used CSS, HTML and JavaScript in developing the project. The calculator does the four basic arithmetic operations which include: Addition, Subtraction, Division and Multiplication. Check it out!